Feat: PR쪽수정 및 코드정리
